% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_sectors.R
\name{attribute_sectors}
\alias{attribute_sectors}
\title{Determine sectors for each point based on trigonometry}
\usage{
attribute_sectors(
  dataset,
  nb_sectors = 8,
  centroid = c(-75.67534, 40.41524),
  rotation = 1
)
}
\arguments{
\item{dataset}{A dataset to be processed}

\item{nb_sectors}{Number of sectors to divide the space in (default: 8)}

\item{centroid}{Coordinates of the centroid to center the circle (lat, long; default: -75.675340, 40.415240)}

\item{rotation}{Number of rotations of the grid wanted (default: 1)}
}
\value{
The same table as \code{dataset} with an additional column named
\code{sectors} and containing the sector number for each row
}
\description{
A full circle is 2*pi, the sector of each point is determined
by atan2(y,x), the angle of the point relative to the horizontal line
of the introduction site
}
\examples{
new_dataset <- attribute_sectors(dataset)
}
