% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefy_groups.R
\name{rarefy_groups}
\alias{rarefy_groups}
\title{Rarefy the jump dataset}
\usage{
rarefy_groups(Jump_groups)
}
\arguments{
\item{Jump_groups}{A dataset to be processed}
}
\value{
A table containing the rarefied dataset
}
\description{
Build the rarefied dataset by keeping only the centroid of each cluster of jumps
}
\examples{
new_dataset <- rarefy_groups(dataset)

}
