% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testNmkdir.R
\name{testNmkdir}
\alias{testNmkdir}
\title{Creates a folder if it does not exist yet}
\usage{
testNmkdir(folder_name)
}
\arguments{
\item{folder_name}{Name of the folder to be tested and created}
}
\value{
Message indicating whether a folder was created or already existed
}
\description{
Creates a folder if it does not exist yet
}
\examples{
testNmkdir("figures")
}
